Add patch: Use /bin/bash to invoke dracut scripts.
The system shell on most Debian and Ubuntu systems is dash, so change
the whack-bang in the Dracut scripts from /bin/sh to /bin/bash.

The `printf "\x$DD\x$CC\x$BB\x$AA" >$TMP` line is problematic because
dash builtin does not implement the hex format.

The dracut/ tree needs more testing and tweaking for the older
dracut-005 package that is in Debian Squeeze and Ubuntu Oneiric.