Move debian/patches into a separate branch.
Using git to automatically rediff the patches is easier than using
`quilt refresh` and manually resolving conflicts, especially because
most submissions for this project come from git remotes.

It is also faster to pull and later discard experimental topic branches.

This also keeps the packaging master history clean and concise, and avoids
accidentally conflating the upstream master. Reverting a commit that changed
something outside of the debian/ overlay is ugly.