Generate META from debian/changelog.
The META file contains a version number that is baked into the kernel module
and appears in the dmesg. Most users report bugs using this information.

At build time, substitute the downstream version in the debian/changelog file
for the upstream version in the META file. This obsoletes the volatile-version
patch.

Closes: dajhorn/pkg-zfs#45