Ensure /etc/zfs/ is present on the system
If /etc/zfs/ isn't present or isn't a directory, zpool.cache
won't be created when importing a pool. Also we need to make
sure when users purges the package zpool.cache is removed as
well.

zpool.cache does not need to be kept on user's system when
the package is purged since information in it is only some
status about the host and mounts. The only direct harm of
not having it is that users need to import the pool by hand
and specify -f parameter.