Cherry picks for root zpool with dracut (Closes: #849969)
After upgrading to systemd 230 as included with Debian stretch, systems
that build their initramfs with dracut and have their root partition on
a zpool fail to boot:

On boot, /usr/lib/dracut/modules.d/98dracut-systemd/rootfs-generator.sh
generates a systemd unit to mount the "root=" kernel parameter, which
for ZFS may look like "zfs:pool/dataset" or "zfs:AUTO", a format not
understood by mount(8).  As a result systemd switches to emergency mode.

A fix was committed upstream in April 2016 with a5a370227eb0 ("fix
booting via dracut generated initramfs"), but never made it into the
0.6.5.* point releases, presumably because they do not pick up fixes
in the contrib/ directory.

Signed-off-by: Aron Xu <aron@debian.org>