Fix installation path of systemd files
The patch makes use of the --with-systemdunitdir configure parameter
which we do not set and which incorrectly defaults to
/usr/lib/systemd/system/.  (Debian uses /lib/systemd/system/.)  Amend
debian/rules to set that parameter. This obviates the need to specify
the /lib/systemd/system/ path in debian/zfs-zed.install and
debian/zfsutils-linux.install, so fix those up while at it.

Signed-off-by: Aron Xu <aron@debian.org>